+++ src/main/java/at/fhv/se/smartmeter/application/port/inbound/meter/CreateMeterUseCase.java
@@ -1,5 +1,5 @@
 package at.fhv.se.smartmeter.application.port.inbound.meter;
 
 public interface CreateMeterUseCase {
-    int createMeter(int serialNumber, String manufacturer);
+    long createMeter(int serialNumber, String manufacturer);
 }
